+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [3.1.6-preview.1] - 2020-04-08
+* Add support for game view eye rendering.
+* Arm32 support
+* Apply frame settings to all cameras and not just the first stereo camera.
+* Add access to Simulation and Render Holographic Frames.
+  
 ## [3.1.5] - 2020-04-08
 *  Back management dependency down to 3.0.6.